requirements.txt
pytest~=8.3.4
Flask~=3.1.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.37